SSP-3,4 added basic weapon system with assault rifle

+++ Assets/GameData/Systems/BasicUnit/AliveUnit.cs
@@ -8,5 +8,5 @@ public abstract class AliveUnit : MonoBehaviour, IDamageble
     [HideInInspector] public float Armour;
 
     public virtual void Die() {}
-    public virtual void TakeDamage(int damagePoints) {} // Add damage type and add damage-multiplyer for enemy-stats
+    public virtual void TakeDamage(float damagePoints) {} // Add damage type and add damage-multiplyer for enemy-stats
 }
