+++ Assets/GameData/Systems/BasicUnit/IDamageble.cs
@@ -1,5 +1,5 @@
 public interface IDamageble
 {
-    public void TakeDamage(int damagePoints);
+    public void TakeDamage(float damagePoints);
     public void Die();
 }
